refactor: Replace StoryGenerator with IStoryGenerator interface and implement Azure Vision API integration
