+++ src/domain/IStoryGenerator.ts
@@ -1,0 +1,4 @@
+export abstract class IStoryGenerator{
+  abstract generate(imageUrl: string): Promise<string>;
+  abstract getClient(): any;
+}
